refactor: simplify prospect screens
