+++ packages/apps/crm/src/components/organisms/Prospect/ProspectBottom/ProspectBottom.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import {EditButton} from '../../../organisms';
+
+const ProspectBottom = ({idProspect, navigation}) => {
+  return (
+    <EditButton
+      onPress={() =>
+        navigation.navigate('ProspectFormScreen', {
+          idProspect: idProspect,
+        })
+      }
+    />
+  );
+};
+
+export default ProspectBottom;
