+++ packages/apps/crm/src/components/organisms/Prospect/index.js
@@ -1,0 +1,3 @@
+export {default as ProspectBottom} from './ProspectBottom/ProspectBottom';
+export {default as ProspectDropdownCards} from './ProspectDropdownCards/ProspectDropdownCards';
+export {default as ProspectHeader} from './ProspectHeader/ProspectHeader';
