+++ packages/apps/crm/src/screens/prospect/ProspectDetailsScreen.js
@@ -1,19 +1,20 @@
 import React, {useEffect} from 'react';
 import {ScrollView} from 'react-native';
-import {Screen, HeaderContainer} from '@axelor/aos-mobile-ui';
+import {Screen, HeaderContainer, NotesCard} from '@axelor/aos-mobile-ui';
 import {
   useSelector,
   HeaderOptionsMenu,
   useDispatch,
+  useTranslator,
 } from '@axelor/aos-mobile-core';
 import {ProspectBody, ProspectBottom, ProspectHeader} from '../../components';
-import {searchContactById} from '../../features/contactSlice';
-import {fetchPartnerEventById} from '../../features/eventSlice';
 import {fetchProspectById} from '../../features/prospectSlice';
 
 const ProspectDetailsScreen = ({navigation, route}) => {
   const idProspect = route.params.idProspect;
+  const I18n = useTranslator();
   const dispatch = useDispatch();
+
   const {mobileSettings} = useSelector(state => state.config);
   const {prospect} = useSelector(state => state.prospect);
 
@@ -35,15 +36,6 @@ const ProspectDetailsScreen = ({navigation, route}) => {
     dispatch(fetchProspectById({partnerId: idProspect}));
   }, [dispatch, idProspect]);
 
-  useEffect(() => {
-    const idList = prospect.contactPartnerSet?.map(item => item.id);
-    dispatch(searchContactById(idList));
-  }, [dispatch, prospect.contactPartnerSet]);
-
-  useEffect(() => {
-    dispatch(fetchPartnerEventById(idProspect));
-  }, [dispatch, idProspect]);
-
   return (
     <Screen removeSpaceOnTop={true}>
       <HeaderContainer
@@ -51,6 +43,7 @@ const ProspectDetailsScreen = ({navigation, route}) => {
         fixedItems={<ProspectHeader />}
       />
       <ScrollView>
+        <NotesCard title={I18n.t('Crm_Notes')} data={prospect.description} />
         <ProspectBody navigation={navigation} />
       </ScrollView>
       <ProspectBottom idProspect={idProspect} navigation={navigation} />
